Move prompt snapshot closure into warp_terminal
Keep terminal model prompt persistence independent of app context-chip and settings modules by colocating the serialized data closure in warp_terminal. Retain app-only generators and GUI conversion behind narrow adapters.
