+++ app/src/context_chips/renderer_tests.rs
@@ -2,7 +2,9 @@ use pathfinder_color::ColorU;
 use warpui::fonts::Properties;
 
 use super::{Renderer, RendererStyles};
-use crate::context_chips::{ChipAvailability, ChipDisabledReason, ContextChipKind};
+use crate::context_chips::{
+    ChipAvailability, ChipDisabledReason, ContextChipKind, ContextChipKindAppExt,
+};
 
 #[test]
 fn test_constructor_availability_updates_disabled_state_and_tooltip_override() {
