+++ app/src/terminal/model/block.rs
@@ -23,6 +23,7 @@ use warp_core::command::ExitCode;
 use warp_core::features::FeatureFlag;
 use warp_errors::report_error;
 use warp_terminal::block_banner::WithinBlockBanner;
+use warp_terminal::context_chips::PromptSnapshot;
 use warp_terminal::model::grid::Dimensions as _;
 use warp_terminal::model::{KeyboardModes, KeyboardModesApplyBehavior};
 use warp_util::lazy::Lazy;
@@ -43,7 +44,6 @@ use super::secrets::RespectObfuscatedSecrets;
 use super::selection::ScrollDelta;
 use super::session::{Sessions, command_executor};
 use crate::ai::agent::redaction::redact_secrets;
-use crate::context_chips::prompt_snapshot::PromptSnapshot;
 use crate::server::ids::SyncId;
 use crate::terminal::block_filter::BlockFilterQuery;
 use crate::terminal::block_list_element::GridType;
